Implement gibbs, icm, bld_appr parameter checks
Since all BLD solvers take the same parameters, parameter validation for
all these functions are moved into a single function.

+++ include/allocation_model/sampling.hpp
@@ -59,7 +59,7 @@ class SampleOnesComputer {
     // computation variables
   private:
     vector_t X_cumsum;
-    size_t X_cols;
+    long X_cols;
     double X_sum;
     util::gsl_rng_wrapper rnd_gen;
 };
